Forked paths.hpp off utils.hpp.
