+++ src/clustering/administration/persist/file.hpp
@@ -8,9 +8,11 @@
 #include "concurrency/rwlock.hpp"
 #include "serializer/types.hpp"
 
+class base_path_t;
 class io_backender_t;
 class merger_serializer_t;
 class filepath_file_opener_t;
+class serializer_filepath_t;
 
 class file_in_use_exc_t : public std::exception {
 public:
