+++ src/concurrency/queue/disk_backed_queue_wrapper.hpp
@@ -11,6 +11,7 @@
 #include "containers/archive/archive.hpp"
 #include "containers/archive/vector_stream.hpp"
 #include "containers/disk_backed_queue.hpp"
+#include "paths.hpp"
 
 /* `disk_backed_queue_t` can't be used directly as a `passive_producer_t`
 because its `pop()` method can sometimes block, and `passive_producer_t`'s
